Actually use SHA-1 implementation in library.
Prior to this, things only worked because on some platforms the SHA1
implementation embedded inside of MoarVM resolved the missing symbols
(note that the library declared SHA1_Init, but called SHA1Init). With
this fix, the SHA1 implementation within this library is really used,
as was surely intended in the first place. Also tidied up the exports
of the library, so it only exports the one symbol we native call from
the Perl 6 code.

With this, the library now works on Windows and builds without any
warnings on Linux.