Do not mutate input buffer when computing the SHA1
In some cases, the input that was to be SHA1'd would be modified, which
was something of a problem if you were, for example, hashing a buffer
and then writing it to a file: the content to be written to the file
was no longer what it originally was! This only seemed to happen if the
buffer was long enough; the case in the added test triggered the
now-fixed bug.